requirements.txt
fastapi==0.115.0
openai==0.28.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
sse-starlette==2.1.3
streamlit==1.38.0
streamlit-audiorec==0.1.3
streamlit-webrtc==0.47.9
uvicorn==0.30.6
